refactor(model-provider): move test files to tests directory
- Move all test files from src/ to tests/ directory
- Update vitest config to look for tests in tests/ directory
- Fix import paths in test files
