+++ multimodal/tarko/model-provider/vitest.config.mts
@@ -7,6 +7,6 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     environment: 'node',
-    include: ['**/*.test.ts'],
+    include: ['tests/**/*.test.ts'],
   },
 });
